+++ modules/core/settings/src/Form/FarmSettingsFarmInfoForm.php
@@ -2,12 +2,9 @@
 
 namespace Drupal\farm_settings\Form;
 
-use Drupal\Core\Config\ConfigFactoryInterface;
 use Drupal\Core\Datetime\TimeZoneFormHelper;
 use Drupal\Core\Form\ConfigFormBase;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\Locale\CountryManagerInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Form for configuring basic farm info.
@@ -15,36 +12,6 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * @ingroup farm
  */
 class FarmSettingsFarmInfoForm extends ConfigFormBase {
-
-  /**
-   * The country manager.
-   *
-   * @var \Drupal\Core\Locale\CountryManagerInterface
-   */
-  protected $countryManager;
-
-  /**
-   * Constructs a RegionalForm object.
-   *
-   * @param \Drupal\Core\Config\ConfigFactoryInterface $config_factory
-   *   The factory for configuration objects.
-   * @param \Drupal\Core\Locale\CountryManagerInterface $country_manager
-   *   The country manager.
-   */
-  public function __construct(ConfigFactoryInterface $config_factory, CountryManagerInterface $country_manager) {
-    parent::__construct($config_factory);
-    $this->countryManager = $country_manager;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container) {
-    return new static(
-      $container->get('config.factory'),
-      $container->get('country_manager')
-    );
-  }
 
   /**
    * {@inheritdoc}
@@ -86,19 +53,6 @@ class FarmSettingsFarmInfoForm extends ConfigFormBase {
     // Get the system.date config.
     $system_date = $this->config('system.date');
 
-    // Get countries.
-    $countries = $this->countryManager->getList();
-
-    // Default country select.
-    $form['default_country'] = [
-      '#type' => 'select',
-      '#title' => $this->t('Default country'),
-      '#empty_value' => '',
-      '#default_value' => $system_date->get('country.default'),
-      '#options' => $countries,
-      '#attributes' => ['class' => ['country-detect']],
-    ];
-
     // Get list of timezones.
     $timezones = TimeZoneFormHelper::getOptionsList();
 
@@ -135,16 +89,12 @@ class FarmSettingsFarmInfoForm extends ConfigFormBase {
       ->set('name', $site_name)
       ->save();
 
-    // Get the submitted country.
-    $default_country = $form_state->getValue('default_country');
-
     // Get the submitted timezone.
     $default_timezone = $form_state->getValue('default_timezone');
 
     // Update system.date config.
     $this->configFactory->getEditable('system.date')
       ->set('timezone.default', $default_timezone)
-      ->set('country.default', $default_country)
       ->save();
 
     // Display message from parent submitForm.